another bolt bucket and config changes
Added another bucket in the bolt schema so that auth leases can be
restored before secret leases, since auth leases are another place
that tokens can hide (and are necessary for restoring secret
leases). This all seems to work.

Also trying out a change in the snapshot config. Replaced `export`
with `remove_after_import`. The idea is, if there's a bolt db file at
the path specified in the config, it should always attempt to restore
from it. If the user doesn't want to restore, remove the file. This
accommodates the long running vm mode. The `remove_after_import`
option will remove the db file after restoring, if the user doesn't
want persistence hanging around (the init and sidecar scenario).